fix: #425 ensure accept/reject port type is a set
If a session is loaded that has an accept_port_types or reject_port_types
field, when the json file is de-serialized, what should be a set
internally becomes a list.

This commit ensures that the data in the model is always a set,
making conversion each time.

(cherry picked from commit e0dbd9fac1ef97148b1c9c6570d6dba11181f604)